refactor: Add eslint rule for import
In general I don't think any files outside of stacks-hierarchy
should import anything from stacks-hierarchy. For now just turn
this on for the components folder.
