+++ src/components/location-icon/index.ts
@@ -1,1 +1,2 @@
 export {LocationIcon} from './LocationIcon';
+export {getVenueIconTypes} from './utils';
