+++ src/stacks-hierarchy/Root_LocationSearchByTextScreen/components/LocationResults.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {TouchableOpacity, View} from 'react-native';
 import {StyleSheet} from '@atb/theme';
-import {LocationIcon} from '@atb/components/location-icon';
+import {getVenueIconTypes, LocationIcon} from '@atb/components/location-icon';
 import {insets} from '@atb/utils/insets';
 import {LocationSearchResultType} from '../types';
 import {FavoriteIcon} from '@atb/favorites';
@@ -12,7 +12,6 @@ import {
   TranslateFunction,
   useTranslation,
 } from '@atb/translations';
-import {getVenueIconTypes} from '../utils';
 import {SearchLocation} from '@atb/favorites';
 
 type Props = {
