+++ src/stacks-hierarchy/Root_LocationSearchByTextScreen/index.ts
@@ -3,4 +3,3 @@ export type {Root_LocationSearchByTextScreenParams} from './navigation-types';
 export {useLocationSearchValue, useOnlySingleLocation} from './hooks';
 export {LocationSearchContent} from './components/LocationSearchContent';
 export type {SelectableLocationType} from './types';
-export {getVenueIconTypes} from './utils';
